Add files via upload
This is our source file of the program. We have already achieved simulating the whole process and appropriately random experiment data. We also designed UI for most of the process. However, we haven't connect the UI to the source file, which made the program currently unable to be used as a perfect teaching tool. We will continuously optimize this program. This file will be available to everyone. We welcome everyone interested to give feedbacks.